refactor: Move `stream` functions from `speakers.py` to `pipes.py`
